[REF] po_hooks: Use class variables instead of dictionary
Instance variables were previously stored inside a dictionary
(which was an instance variable itself). This added
boilerplate to the code (had to type out the dict name to
access a variable) and prevented IDEs and language servers from
performing typing inference.

Therefore the code has been refactored to store this dictionary
keys as instance variables directly.