[REF] file-not-used: Skip option using "oca_data_manual" in the manifest file
Sometimes you want to load data files using a python script

e.g. post_init_hook

The manifest will not have these files but in a valid way

But this check is raising false red

So, we have added an option to use consider it as valid

Adding in the __manifest__.py a custom key called  "oca_data_manual"

e.g.

    {
        ...
        "data": ["data/normal_data.xml"],
        "oca_data_manual": ["data/manual_data.xml"],
    }