[REM] xml-dangerous-filter-wo-user: Remove outdated check for dangerous filters without user
The check `xml-dangerous-filter-wo-user` has been removed as it is no longer relevant
after the following fix already merged:
 - https://github.com/odoo/odoo/pull/8218

This fix was already merged for old Odoo versions and this checks could be removed
since that the new Odoo versions do not have this issue and the old versions are not using this tool

The issue born from Odoo 8.0 still reproduced in 10.0
and already fixed for 12.0+ versions.

So, this check did the work to avoid using dangerous filters without user context in XML files during a few years,

but now it is safe to remove it from the pre-commit hooks.