[REM] xml-dangerous-filter-wo-user: Remove outdated check for dangerous filters without user (#139)
The check `xml-dangerous-filter-wo-user` has been removed as it is no longer relevant
after the following fix was already merged:
 - https://github.com/odoo/odoo/pull/8218

The Odoo issue was born from Odoo 8.0, and it is still reproduced in 10.0
It was already fixed for 12.0+ versions.

So, this check did the work to avoid using dangerous filters without user context in XML files for a few years,
But now it is safe to remove it from the pre-commit hooks since that 12.0 version is too old now
And the current projects using old Odoo 11.0- versions were not using odoo-pre-commit-hooks;
they were using the old tool, where this check is even enabled.